Split instrument discovery and import #530
